refactor registry
